PrestoDB 0.227 Alluxio connector
No logical code changes were required for to add the Alluxio
connector on top of the 0.227 tag. The only changes required were
related to changing some of the com.facebook.airlift imports to
io.airlift. Additionally, on the master branch of prestoDB (at
the time of this commit) the metastore was factored out into its
own maven module, so I had to incorporate these files back into
the `presto-hive` module, similar to prestoSQL