Implement Iceberg changelog scanning
This commit exposes a new system table called $changelog that
allows users to return a table pertaining to info about when
which snapshots specific records were added or deleted.

The schema of the changelog table is as follows:
| primary_key: T | operation: VARCHAR | ordinal: bigint | snapshotid: bigint |

To use this feature, the only requirement is that users need to specify
the name of a column in the table properties as the "primary key". The
value of the key will be exposes in the "primary_key" column of the
changelog table. The type of the column will be the same type as in
the underlying table.

The current beahvior in this commit only returns the results of the
changelog between the current and previous snapshot of the table being
queried. You can use this information to determine when a record has
been inserted or deleted.

SELECT committed_at, change.operation, primary_key, ordinal
FROM "orders$snapshots" as snap
JOIN "orders$changelog" as change
ON change.snapshotid = snap.snapshot_id;

               committed_at                 | operation | primary_key | ordinal
---------------------------------------------+-----------+-------------+---------
 2023-07-19 13:03:03.207 America/Los_Angeles | INSERT    |        1956 |       0
 2023-07-19 13:03:03.207 America/Los_Angeles | INSERT    |       28196 |       0
 2023-07-19 13:03:03.207 America/Los_Angeles | INSERT    |       27873 |       0

At a high-level the implementation strategy needed to expose information
from Iceberg's IncrementalChangelogScan tasks which differs from a
normal TableScan. The plan tasks for IncrementalScans include the
information on operation type, ordinal, and snapshotId which are are not
contained in the normal table scan plan.

When the connector metadata queries for a changelog table type, it
returns a schema with a static 4 columns, where the first 3 are static
VARCHAR, BIGINT, BIGINT, and the final is dependent upon the user-
defined primary key property.

To accomplish this the following changes were made
- a new primary key table property was added
- logic to generate the changelog table schema
- a new piece of info on iceberg splits containing the original table's
column handles which is used during split sourcing
- a new page source which constructs the correct format page based on
the extracted primary key.
- finally, some additional refactoring of the IcebergNativeMetadata and
IcebergHiveMetadata classes to consolidate certain bits of logic for the
ConnectorMetadata

And of course, this is preliminary work. There is still some more
optimization that could be done.

TODOs
- Add tests
- support table predicate
- support more than just parent ancestor changelog
- filter on a predicate when generating the incremental change log scan
- optimize to scan only for the primary key rather than the whole row