Prevent scanning additional data for changelog
Since the changelog only surfaces the primary key, we don't need the
delegate page source to scan all of the other data. Previously the
implementation would scan for all of the rows since we always passed the
full set of column handles.

This change filters the set of column handles passed down to just the
primary key which should limit the amount of data scanned when querying
the changelog table.