Implement basic coalesce function
This is a POC for a coalesce function which combines
inputs in order to determine the  final set of values for
the changes in an iceberg table's changelog.

The current method is to implement the coalesce as an
aggregation function which is then able to be called as
a window function. The reason to implement it in this
way is because the coalescing of rows acts as an
aggregation function. Regular window function
implementations are limited to outputting a single value
per input row.

Since this is a POC, there are a number of TODOs to polish
this into a usable state

1. Move the function implementation into the iceberg
connector. This requires a bit of interface boilerplate.
2. add actual row data to the changelog
3. implement the grouped state type for the aggregation
function state