Get row data in the changelog
This change adds a new column "rowdata" to the changelog which
contains all of the row information for each changelog entry.

The inclusion of row data will increase the amount of data scanned, but
allows us to retrieve the entire set of data required to insert rows
into the samples table after coalescing the changelog.