Move changelog function to iceberg module
Since the changelog only applies to iceberg tables, this function should
only be loaded with the iceberg connector.

In addition, needed to add support for functions which don't include
any of the @AggregationFunction or @ScalarFunction annotations
(generally these are functions which accept generic arguments). So,
I've implemented support for a new @FunctionInstance annotation
which should grab and register these functions as well without
mucking too much with the SPI interface

Also, we needed to add the dependency on presto-main in order to
allow the changelog function to extend from SqlAggregationFunction