Add a function to aggregate the changelog (#10)
* Implement basic coalesce function

This is a POC for a coalesce function which combines
inputs in order to determine the  final set of values for
the changes in an iceberg table's changelog.

The current method is to implement the coalesce as an
aggregation function which is then able to be called as
a window function. The reason to implement it in this
way is because the coalescing of rows acts as an
aggregation function. Regular window function
implementations are limited to outputting a single value
per input row.

Since this is a POC, there are a number of TODOs to polish
this into a usable state

1. Move the function implementation into the iceberg
connector. This requires a bit of interface boilerplate.
2. add actual row data to the changelog
3. implement the grouped state type for the aggregation
function state

* Get row data in the changelog

This change adds a new column "rowdata" to the changelog which
contains all of the row information for each changelog entry.

The inclusion of row data will increase the amount of data scanned, but
allows us to retrieve the entire set of data required to insert rows
into the samples table after coalescing the changelog.

* Fix changelog tests after new rowdata column addition

* Move changelog function to iceberg module

Since the changelog only applies to iceberg tables, this function should
only be loaded with the iceberg connector.

In addition, needed to add support for functions which don't include
any of the @AggregationFunction or @ScalarFunction annotations
(generally these are functions which accept generic arguments). So,
I've implemented support for a new @FunctionInstance annotation
which should grab and register these functions as well without
mucking too much with the SPI interface

Also, we needed to add the dependency on presto-main in order to
allow the changelog function to extend from SqlAggregationFunction

* refactor function name to apply_changelog