Add arguments for merging into an initial sample
Two new arguments are added to the signature of the reservoir_sample
function which indicate an array of existing samples and the total
number of samples processed to generate that sample.

Additionally, the output signature now uses named fields in order to be
able to select the seen count and samples individually.

I changed the structure of the ReservoirampleState and
ReservoirSample classes to not use a boolean to track "isEmpty" and
use a boolean function to calculate it. In place of using the "isEmpty"
variables, we track the current "maxSampleSize" versus the actual size,
"samples.size()". This is the main reasoning for switching to an
ArrayList. There should be no performance impact as long as the
ArrayList is always initialized with the expected max count.

Lastly, some optimizations were made to correctly pass the expected
block sizes to a few of the block builder methods.