Add HISTOGRAM to column statistic types
This change adds the `HISTOGRAM` type to the enum. In addition
a new field "functionName" is added to the ColumnStatisticMetadata
class.

The change to ColumnStatisticMetadata is in preparation to allow
connectors to override the function used to execute the statistics
aggregations. Mainly it is to support connectors that have differing
underlying histogram representations. For example, Hive natively
generates KLL sketches from Apache Datasketches, Spark has it's own
format. Some other connectors like mysql use a JSON format. If we intend
to support other connectors this change is necessary.

Allowing the connector to override the function will let us generate the
histogram in the format needed by the connector in order to store the
histogram in the connector-specific catalog. Further down the line,
we'll read these histograms and connectors can supply specific
implementations to the CBO.