Add histograms for optimizer cost calculation
This commits provides two critical changes:

1. Adds a new enum value to ColumnStatisticType: Histogram.
2. Utilizes the new histograms in optimizer's cost calculations

Additional details below:

With this change, a new column statistic type for histograms is
introduced. In addition, a new SPI class `ConnectorHistogram`
is also introduced. This interface is designed to be able to be
implemented by either the connectors or in the main presto
codebase. This should allow connectors to return utilize
histogram statistics in any format regardless of the source.

The API is straightforward and includes 4 methods.

- cumulativeProbability(double value): -> CDF function
- inverseCumulativeProbability(double probability) -> inverse CDF function
- cumulativeDistinctValues(double value) -> similar to CDF, but for counting distinct values
- distinctValues() -> the total distinct values in the distribution

A reference implementation is provided inside of
UniformDistributionHistogram. This implementation results in the same
logic and same plans as the previous cost-based calculations. The math
ends up being the same, but just utilizing the histogram API.

Additionally, to propagate statistics further up into a plan, another
implementation of histograms is provided inside of
DomainConstrainedHistogram. This class is used to bound a source
histogram with a given domain as additional filters may be applied
further up in the plan.

Previously all cost calculations were performed inside of the
ComparisonStatsCalculator using logic from the StatisticRange class
to calculate the filter factors of overlapping and intersecting ranges

Instead, this change introduces cost calculation using the new histogram
model and API. The core of the filter proportion calculation logic
exists in the new HistogramCalculator utility class.

If the underlying Histogram implementation is swapped from the
UniformDistributionHistogram, then the stats calculator should
calculate the costs using the histogram information correctly.

I verified through additional unit tests and the existing testing
infrastructure that the implementation here results in the same cost and
filter proportion calculations as the previous logic.