Add discovery and memory functionality
Adds the ability to announce to the discovery server that the server is live. This includes
a new task spawned at the start of the process which sends a hard-coded service
announcement message every 20 seconds or so. Eventually will probably want to
configure this. The discovery URI is currently hard-coded.

In addition, this commit also implements the basic /v1/memory APIs such that the
coordinator sees this node as a valid node to schedule tasks to. Running queries on
the coordinator will now attempt to schedule tasks via the task control plane API
which is currently unimplemented.