Update protocol for better json handling
all connector types with @type will eventually need custom serde.
For now, we serialize them into class ID and raw JSON which we'll punt to parse later on. For now we will be able to successfully handle at least
the concrete types.

Additionally, adds some params to the language map to support serde
renames. Also, moved the protocol to `resources.rs` and into a new mod which contains additional impls (e.g. hash and serialization) for
types with special cases.