Working execution pipelines and block deserialization
Trick for getting pipelines to execute properly was to drop the
sender once the pipeline completes. The finished message can be
received out of order from the input pages, so just keep looping
until the buffer has been flushed and then None will come down
on the receiver which indicates it's time to finish executing.