Add spilling to memory connector
There are a number of changes in this commit which significantly alter
the behavior of the connector.

1. Splits per node can now be configured through a session property.
2. Pages are written to a spillable store. The memory limit and spill
    file size limit configs are added to the MemoryConfig class. When
    the memory limit is exceeded, pages are spilled to disk. After
    exceeding the spill file limit, new spill files are created. They
    are not in Pagefile format. It is just a stream of SerializedPage
3. The memory.max-data-per-node config now takes on a new role which
    limits the maximum amount of in-memory data that can be stored
    before forcing flushes to disk. Reaching the limit will flush all
    existing in-memory tables to disk.

Some caveats:

- Parallel reads (e.g. more than 1 split per node) seems to degrade
performance. This can be improved in future designs to more efficiently
read subsets of the iterator. This is the main reason for adding the new
session configuration
- Concurrent read/writes are not consistent. Reading the tuple store
while writes are occurring may result in missing some pages while
iterating. This can occur while iterating over the in-memory page queue
while a spilling action is initiated.