Fix race with CTE split scheduling
This change fixes a race condition which caused the memory split source
to underreport the number of splits for a particular table.

When a TableCommit operator executes it calls the ConnectorMetadata's
`finishInsert` method. In the memory connector, finishInsert on a table
updates a table's list of data fragments. The data fragments are just a
simple set of host->row count mappings. However, an updated list of
fragments is required in order to properly generate splits for each
worker.

With CTEs enabled, during the query scheduling process, we call
out to the split manager to get the split source even before the CTE
table write is complete. The call to `getNextBatch` on the split source
doesn't occur until scheduling time. In the memory connector, we compute
all splits at SplitSource creation time. This led to a bug where if CTE
table writers didn't commit early enough that the readers would get split
sources with splits missing from some workers.

In order to fix the bug, we needed to delay split calculation until
the time at which SplitSource#getNextBatch is called. To achieve this we
change the list of MemoryTableFragment to an immutable view of the
underlying fragment list. This allows for the computation of the split
source to view the newly updated splits. Then, we also delay the
computation of the splits using a memoizedSupplier until the first call
of `getNextBatch` using the MemoryLazySplitSource.

As long as we guarantee TableCommit(s) finish before the first call to
`getNextBatch`, then this solution should be viable. It seems to