Force MySQL connection to use TLSv1.2
Newer versions of JDK have TLS 1.0 and 1.1 disabled by default. The JDK 8 version we were using was never updated to a newer one, so it continued to use these TLS versions

The right fix is to update ConnectorJ (mysql-connector-java) and the MySQL versions we use 8.x. This should fix the defaults to sane TLS values.

For now, we force TLSv1.2 to maintain parity with how we ran tests before. Disabling SSL/TLS with `useSSL=false` is also a valid setup for this test

Refs -
https://community.streamsets.com/common-issues-47/mysql-jdbc-error-javax-net-ssl-sslhandshakeexception-no-appropriate-protocol-112
https://bugs.openjdk.org/browse/JDK-8256490