Split maven modules with dependency on jetty http server
This commit splits presto-main into presto-main
and presto-main-base. presto-main-base is now the "main"
module which builds the PrestoServer binary which includes
the airlift http-server dependency. presto-main
is all of the internal classes required to run presto and
does not depend on jetty.

This is in preparation to upgrade to Java 17 within Presto.
The latest supported versions of Jetty require Java 17. This
change allows us to maintain Presto-On-Spark compatibility
with Java 8 as long as we keep the presto-main module compiling
to java 8.

Eventually these modules may be recombined after all PoS users
have moved to Java 17 runtime.