[Iceberg] deprecate some table property names
This commit introduces the deprecation of a few tables property names
in favor of the iceberg library's naming scheme. It is to help reduce
confusion for users as to which table properties in Iceberg documentation
map to presto.

Users will get a warning on any queries which set table properties such
as CREATE TABLE and ALTER TABLE .. SET PROPERTIES statements.

Refer to IcebergTableProperties.java or iceberg.rst for the list of
deprecated names and their corresponding mapping.